Change return types to uint16_t and update warnings

+++ examples/word_sequence.c
@@ -1,3 +1,10 @@
+/*
+ * word_sequence.c
+ *
+ *  Created on: Oct 14, 2025
+ *      Author: luong
+ */
+
 // In word_sequence.c
 #include "word_sequence.h"
 #include <stdio.h>
@@ -6,8 +13,10 @@
  * Initialize the word sequence filter state.
  * @param state: Pointer to the filter's state structure.
  */
-void word_sequence_init(word_sequence_state_t *state) {
-	if (!state) {
+void word_sequence_init(word_sequence_state_t *state)
+{
+	if (!state)
+	{
 		return;
 	}
 	state->s1 = 0;
@@ -20,8 +29,10 @@ void word_sequence_init(word_sequence_state_t *state) {
  * @param current_sy: The current input symbol (s0).
  * @return: The calculated word value.
  */
-int32_t word_sequence_update(word_sequence_state_t *state, uint8_t current_sy) {
-	if (!state) {
+uint16_t word_sequence_update(word_sequence_state_t *state, uint8_t current_sy)
+{
+	if (!state)
+	{
 		return 0; // Return a default value on error
 	}
 
@@ -33,7 +44,7 @@ int32_t word_sequence_update(word_sequence_state_t *state, uint8_t current_sy) {
 	uint8_t s2 = state->s2;
 
 	// Create the word value using the current and previous symbols
-	int32_t word_value = (s2 << 8) | (s1 << 4) | s0;
+	uint16_t word_value = (s2 << 8) | (s1 << 4) | s0;
 
 	// Update the state for the next function call
 	// The current s1 becomes the next s2
@@ -43,9 +54,10 @@ int32_t word_sequence_update(word_sequence_state_t *state, uint8_t current_sy) {
 
 	// Clamp the value if it exceeds the known buffer limit from the
 	// shannon_entropy function
-	if (word_value >= 2458) {
+	if (word_value >= 2458)
+	{
 		fprintf(stderr,
-				"word_sequence: Warning - word value %ld exceeds buffer "
+				"word_sequence: Warning - word value %d exceeds buffer "
 				"limit, clamping to 2457\n",
 				word_value);
 		word_value = 2457;
